Refactor jsTree event handling to enable "open all" for Projects
In order to provide functionality to open all Dataset children for a
selected Project the component hierarchy needs to reflect what nodes
of the jsTree that are *open* rather than just those that are
*selected*.  It is easiest to do this via component key at the right
level which will affect mount/unmount status and this is exactly what
this commit does.

The stage is further set to enable the same type of logic for Screens
once we are ready.