Closely mimic thumbnail plugin event handling and dispatch
If we listen to "select_node.jstree" or "changed.jstree" we will both be
smashing the jsTree with events and be smashed by the jsTree by events.
All of which we would end up having to reconcile in our React components
to accurately reflect legitimate change.

The existing thumbnail plugin works around this issue by disabling
jsTree event propogation when making large selections and then by making
use of the "selection_change.ome" event.  As the jsTree accurately
bubbles "selection_change.ome" events up for the purposes of several
other components we can also listen to these exclusively for our own
purposes.