Handle scenarios where we initially have transient state
Under certain conditions when we recieve a "selection_change.ome" event
the jsTree has not yet been loaded.  This seems to be isolated to
scenarios where the effective root node is a Dataset but it is certainly
possible that it is happening during elsewhere.

This ensures that when the resultant component update happens that the
transient state is noticed and the image JSON is brought in sync.