Move configuration from window global
Configuration is now available globally from a module rather than a
series of variables on the window global.  Various cases of
hardcoded URLs have also been cleaned up.

Parade should now be capable of operating completely independently of
what the OMERO.web base URL might be set at.

We might want to use this configuration with one or more React contexts
in the future but for right now it is isolated.